Add GitHub action to amend changlog entries
This adds a GitHub action that automatically adds pr links to any
changelog entries staged in the pull request.

This has the unfortuante side effect of hiding the workflow run in
the GitHub UI, so this action is deliberately run AFTER all the CI
has been run, and only if it passes.